requirements.txt
pipecat-ai[cartesia,openai,silero,deepgram,elevenlabs,daily]
fastapi
uvicorn
python-dotenv
loguru